+++ modules/saxs/include/ChiScore.h
@@ -18,18 +18,18 @@ IMPSAXS_BEGIN_NAMESPACE
 */
 class IMPSAXSEXPORT ChiScore {
  public:
-  Float compute_score(const Profile* exp_profile, const Profile* model_profile,
-                      bool use_offset = false) const;
+  double compute_score(const Profile* exp_profile, const Profile* model_profile,
+                       bool use_offset = false) const;
 
-  Float compute_score(const Profile* exp_profile, const Profile* model_profile,
-                      Float min_q, Float max_q) const;
+  double compute_score(const Profile* exp_profile, const Profile* model_profile,
+                      double min_q, double max_q) const;
 
-  Float compute_scale_factor(const Profile* exp_profile,
-                             const Profile* model_profile,
-                             Float offset = 0.0) const;
+  double compute_scale_factor(const Profile* exp_profile,
+                              const Profile* model_profile,
+                              double offset = 0.0) const;
 
-  Float compute_offset(const Profile* exp_profile,
-                       const Profile* model_profile) const;
+  double compute_offset(const Profile* exp_profile,
+                        const Profile* model_profile) const;
 };
 
 IMPSAXS_END_NAMESPACE
